Implement the plan
This commit implements the changes as per the provided plan.
