+++ src/pages/Profile.tsx
@@ -1,9 +1,11 @@
-import { useState, useEffect } from "react";
+
+import { useState, useEffect, useCallback } from "react";
 import { useNavigate } from "react-router-dom";
-import { DollarSign, Shield, User, FileText, Pi, AlertTriangle, Bell } from "lucide-react";
+import { MiniKit, VerifyCommandInput, VerificationLevel, ISuccessResult } from "@worldcoin/minikit-js";
+import { DollarSign, Shield, User, FileText, Pi, AlertTriangle, Bell, Globe } from "lucide-react";
 import { motion } from "framer-motion";
 import { Header } from "@/components/Header";
-import { useMagnifyWorld } from "@/hooks/useMagnifyWorld";
+import { useMagnifyWorld, Tier } from "@/hooks/useMagnifyWorld";
 import { Card } from "@/components/ui/card";
 import { Alert, AlertTitle, AlertDescription } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
@@ -14,8 +16,10 @@ const Dashboard = () => {
   // hooks
   const navigate = useNavigate();
   const ls_username = localStorage.getItem("ls_username");
-  const ls_wallet = localStorage.getItem("ls_wallet_address");
+  const ls_wallet = localStorage.getItem("ls_wallet_address") || "";
   const { data, isLoading, isError, refetch } = useMagnifyWorld(ls_wallet as `0x${string}`);
+  const [verifying, setVerifying] = useState(false);
+  const [currentTier, setCurrentTier] = useState<Tier | null>(null);
 
   // state
   const nftInfo = data?.nftInfo || { tokenId: null, tier: null };
@@ -23,9 +27,113 @@ const Dashboard = () => {
   const loan = data?.loan;
 
   // Check if the user is verified by ORB device
-
   const isOrbVerified = nftInfo?.tier?.verificationStatus?.verification_level === "orb";
   
+  // Verification levels
+  const verificationLevels = {
+    orb: {
+      tierId: BigInt(2),
+      level: "Orb Scan",
+      icon: Globe,
+      action: "mint-orb-verified-nft",
+      upgradeAction: "upgrade-orb-verified-nft",
+      verification_level: VerificationLevel.Orb,
+    },
+  };
+
+  // Handle verification process
+  const handleVerify = useCallback(async (tier: typeof verificationLevels.orb) => {
+    if (!MiniKit.isInstalled()) {
+      toast({
+        title: "Verification Failed",
+        description: "Please install World App to verify.",
+        variant: "destructive",
+      });
+      return;
+    }
+
+    setVerifying(true);
+    setCurrentTier(tier as unknown as Tier);
+
+    const verificationStatus = {
+      claimAction: tier.action,
+      upgradeAction: tier.upgradeAction,
+      verification_level: tier.verification_level,
+      level: tier.level,
+    };
+
+    const verifyPayload: VerifyCommandInput = {
+      action: verificationStatus.claimAction || verificationStatus.upgradeAction,
+      signal: ls_wallet,
+      verification_level: verificationStatus.verification_level as VerificationLevel,
+    };
+
+    try {
+      const { finalPayload } = await MiniKit.commandsAsync.verify(verifyPayload);
+      if (finalPayload.status === "error") {
+        console.error("Verification failed:", finalPayload);
+
+        if (finalPayload.error_code === "credential_unavailable") {
+          toast({
+            title: "Verification Failed",
+            description: "You are not Orb Verified in the WorldChain App. Please complete Orb verification first.",
+            variant: "destructive",
+          });
+        } else {
+          toast({
+            title: "Verification Failed",
+            description: "Something went wrong. Please try again.",
+            variant: "destructive",
+          });
+        }
+        setVerifying(false);
+        return;
+      }
+
+      // Send proof to backend
+      const response = await fetch(`${BACKEND_URL}/verify`, {
+        method: "POST",
+        body: JSON.stringify({
+          payload: finalPayload as ISuccessResult,
+          action: verificationStatus.claimAction || verificationStatus.upgradeAction,
+          signal: ls_wallet,
+        }),
+      });
+
+      const result = await response.json();
+      if (response.ok) {
+        toast({
+          title: "Verification Successful",
+          description: `You are now ${verificationStatus.level} Verified.`,
+        });
+        refetch();
+      } else {
+        toast({
+          title: "Backend Error",
+          description: result.message || "Something went wrong.",
+          variant: "destructive",
+        });
+        console.error("Backend verification failed:", result);
+      }
+    } catch (error: any) {
+      console.error("Error during verification:", error);
+
+      let errorMessage = "Something went wrong while verifying.";
+
+      if (error?.error_code === "credential_unavailable") {
+        errorMessage = "You are not Orb Verified in the WorldChain App. Please complete Orb verification first.";
+      }
+
+      toast({
+        title: "Verification Failed",
+        description: errorMessage,
+        variant: "destructive",
+      });
+    } finally {
+      setVerifying(false);
+    }
+  }, [ls_wallet, refetch, toast]);
+
   if (isLoading) {
     return (
       <div className="min-h-screen">
@@ -45,10 +153,13 @@ const Dashboard = () => {
   }
 
   if (!isLoading && data) {
+    const nftInfo = data?.nftInfo || { tokenId: null, tier: null };
+    const userTierId = nftInfo?.tier?.tierId || BigInt(0);
+    const isDeviceVerified = nftInfo?.tier?.verificationStatus?.verification_level === "device";
+
     return (
       <div className="min-h-screen bg-background">
         <Header title="Profile" />
-        {/* Header */}
         <div className="max-w-4xl mx-auto space-y-8 px-4 py-6">
           {/* User Profile */}
           <motion.div
@@ -67,6 +178,61 @@ const Dashboard = () => {
             ) : (
               <p className="text-muted-foreground text-center text-lg">Unverified</p>
             )}
+          </motion.div>
+
+          {/* Verification Section */}
+          <motion.div
+            initial={{ opacity: 0, y: 20 }}
+            animate={{ opacity: 1, y: 0 }}
+            className="glass-card p-6 mb-8"
+          >
+            <div className="flex items-center justify-center mb-4">
+              <Shield className="w-12 h-12 text-primary" />
+            </div>
+            <h2 className="text-2xl font-bold text-gradient mb-2 text-center">Verification Level</h2>
+            <p className="text-muted-foreground text-center text-lg mb-6">
+              {isDeviceVerified || nftInfo.tokenId === null ? "Unverified" : `Currently: ${nftInfo.tier?.verificationStatus.level.charAt(0).toUpperCase() + nftInfo.tier?.verificationStatus.level.slice(1).toLowerCase()} Verified`}
+            </p>
+
+            {/* Verification Options */}
+            <div className="grid grid-cols-1 md:grid-cols-1 gap-6">
+              {Object.values(verificationLevels).map((tier) => {
+                const IconComponent = tier.icon;
+
+                return (
+                  <motion.div
+                    key={tier.level}
+                    initial={{ opacity: 0, y: 20 }}
+                    animate={{ opacity: 1, y: 0 }}
+                    transition={{ delay: 0.1 }}
+                    className="glass-card p-6"
+                  >
+                    <IconComponent className="w-12 h-12 mx-auto mb-4 text-primary" />
+                    <h3 className="text-xl font-semibold mb-2 text-center">{tier.level} Verification</h3>
+
+                    <Button
+                      className="w-full"
+                      variant="default"
+                      disabled={
+                        verifying || // Disable while verifying
+                        userTierId > tier.tierId || // User already at a higher tier
+                        tier.verification_level === nftInfo?.tier?.verificationStatus.verification_level
+                      }
+                      onClick={() => handleVerify(tier)}
+                    >
+                      {verifying && currentTier?.tierId === tier.tierId
+                        ? "Verifying..."
+                        : nftInfo.tokenId === null || isDeviceVerified
+                        ? "Claim NFT"
+                        : userTierId > tier.tierId ||
+                          tier.verification_level === nftInfo?.tier?.verificationStatus.verification_level
+                        ? "Already Claimed"
+                        : `Upgrade to ${tier.level}`}
+                    </Button>
+                  </motion.div>
+                );
+              })}
+            </div>
           </motion.div>
 
           {/* Collateral section */}
@@ -112,25 +278,13 @@ const Dashboard = () => {
                 </motion.div>
               </div>
             </motion.div>
-          ) : (
-            <div className="flex-column justify-center items-center h-[calc(100vh-80px)]">
-              <h2 className="text-2xl font-semibold mb-4">You are unverified</h2>
-              <p className="mb-4">
-                To be eligible for a loan, you need to own a specific NFT and be verified by an ORB device. Please upgrade your account to include this NFT.
-              </p>
-              <button
-                onClick={() => navigate("/upgrade-verification")}
-                className="glass-button w-full"
-                disabled={isLoading}
-              >
-                Upgrade Now
-              </button>
-            </div>
-          )}
+          ) : null}
         </div>
       </div>
     );
   }
+
+  return null;
 };
 
 export default Dashboard;
